Build an empty addon on win32
Npm does not support platform sensitive dependencies, and will always
attempt to build syslog even if its not going to be used. The error
messages are disturbing to naive users. Instead, compile syslog into a
an empty addon, so that it builds sucessfully, but won't have an .init()
method. This supports run-time detection of syslog capability, rather
than npm install time build failures.